+++ package.json
@@ -11,8 +11,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "4.0.3",
-    "uuid": "^8.3.2",
-    "web-vitals": "^1.1.1"
+    "uuid": "^8.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
