+++ src/ConstantValues.js
@@ -1,0 +1,53 @@
+export const data =  [
+    {
+      id: "5PXCa6RCG28",
+      title: "8 Camera HACKS in 90 SECONDS!!"
+    },
+    {
+      id: "QHDhSidFhcQ",
+      title: "AFTER EFFECTS BASICS"
+    },
+    {
+      id: "GYud_W7Ou9E",
+      title: "Make your footage look Cinematic FAST! Premiere Pro Tutorial"
+    },
+    {
+      id: "hnMY1u7M90o",
+      title: "What LENS should YOU BUY?!"
+    },
+    {
+      id: "e6HZPmSlS5c",
+      title: "What Frame Rate Should You Be Filming In?"
+    },
+    {
+      id: "JNXmjTmuGNM",
+      title: "10 MISTAKES EVERY BEGINNER FILMMAKER MAKES"
+    },
+    {
+      id: "boeL109kC3A",
+      title: "How I film a cinematic travel video - Sydney BTS and bloopers!"
+    },
+    {
+      id: "r25IWquxe9s",
+      title: "THE BUCKET SHOT"
+    },
+    {
+      id: "CIrowsD7QBk",
+      title: "Exploring Indonesia - The Last Paradise"
+    },
+    {
+      id: "JcinvbS4Fi0",
+      title: "Filming a SPICY Commercial with JUST A TRIPOD! | Behind the Scenes"
+    },
+    {
+      id: "458I2hkaqjk",
+      title: "In Camera Transitions: A Masterclass."
+    }
+  ];
+
+export const ADD_TO_LIKED = "ADD_TO_LIKED";
+export const REMOVE_FROM_LIKED = "REMOVE_FROM_LIKED";
+export const CREATE_PLAYLIST = "CREATE_PLAYLIST";
+export const DELETE_PLAYLIST = "DELETE_PLAYLIST";
+export const ADD_TO_PLAYLIST = "ADD_TO_PLAYLIST";
+export const REMOVE_FROM_PLAYLIST = "REMOVE_FROM_PLAYLIST";
